requirements.txt
numpy
paho-mqtt
PyYAML
scipy
pandas